+++ src/client.ts
@@ -61,6 +61,7 @@ import {
   CreateImportResponse,
   CreateImportURLResponse,
   CustomPermissionOptions,
+  DeactivateUsersOptions,
   DefaultGenerics,
   DeleteCampaignOptions,
   DeleteChannelsResponse,
@@ -120,6 +121,8 @@ import {
   PushProviderUpsertResponse,
   QueryChannelsAPIResponse,
   ReactionResponse,
+  ReactivateUserOptions,
+  ReactivateUsersOptions,
   Recipient,
   RecipientFilters,
   RecipientQueryOptions,
@@ -179,9 +182,7 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
   clientID?: string;
   configs: Configs<StreamChatGenerics>;
   key: string;
-  listeners: {
-    [key: string]: Array<(event: Event<StreamChatGenerics>) => void>;
-  };
+  listeners: Record<string, Array<(event: Event<StreamChatGenerics>) => void>>;
   logger: Logger;
   /**
    * When network is recovered, we re-query the active channels on client. But in single query, you can recover
@@ -212,6 +213,7 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
   insightMetrics: InsightMetrics;
   defaultWSTimeoutWithFallback: number;
   defaultWSTimeout: number;
+  private nextRequestAbortController: AbortController | null = null;
 
   /**
    * Initialize a client
@@ -599,7 +601,10 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
     if (apn_config?.p12_cert) {
       options = {
         ...options,
-        apn_config: { ...apn_config, p12_cert: Buffer.from(apn_config.p12_cert).toString('base64') },
+        apn_config: {
+          ...apn_config,
+          p12_cert: Buffer.from(apn_config.p12_cert).toString('base64'),
+        },
       };
     }
     return await this.patch<APIResponse>(this.baseURL + '/app', options);
@@ -621,7 +626,9 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
    * Revokes all tokens on application level issued before given time
    */
   async revokeTokens(before: Date | string | null) {
-    return await this.updateAppSettings({ revoke_tokens_issued_before: this._normalizeDate(before) });
+    return await this.updateAppSettings({
+      revoke_tokens_issued_before: this._normalizeDate(before),
+    });
   }
 
   /**
@@ -645,7 +652,9 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
     for (const userID of userIDs) {
       users.push({
         id: userID,
-        set: <Partial<UserResponse<StreamChatGenerics>>>{ revoke_tokens_issued_before: before },
+        set: <Partial<UserResponse<StreamChatGenerics>>>{
+          revoke_tokens_issued_before: before,
+        },
       });
     }
 
@@ -755,7 +764,10 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
 
     this.anonymous = true;
     this.userID = randomId();
-    const anonymousUser = { id: this.userID, anon: true } as UserResponse<StreamChatGenerics>;
+    const anonymousUser = {
+      id: this.userID,
+      anon: true,
+    } as UserResponse<StreamChatGenerics>;
 
     this._setToken(anonymousUser, '');
     this._setUser(anonymousUser);
@@ -779,10 +791,12 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
     let response: { access_token: string; user: UserResponse<StreamChatGenerics> } | undefined;
     this.anonymous = true;
     try {
-      response = await this.post<APIResponse & { access_token: string; user: UserResponse<StreamChatGenerics> }>(
-        this.baseURL + '/guest',
-        { user },
-      );
+      response = await this.post<
+        APIResponse & {
+          access_token: string;
+          user: UserResponse<StreamChatGenerics>;
+        }
+      >(this.baseURL + '/guest', { user });
     } catch (e) {
       this.anonymous = false;
       throw e;
@@ -842,11 +856,15 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
     if (!(key in this.listeners)) {
       this.listeners[key] = [];
     }
-    this.logger('info', `Attaching listener for ${key} event`, { tags: ['event', 'client'] });
+    this.logger('info', `Attaching listener for ${key} event`, {
+      tags: ['event', 'client'],
+    });
     this.listeners[key].push(callback);
     return {
       unsubscribe: () => {
-        this.logger('info', `Removing listener for ${key} event`, { tags: ['event', 'client'] });
+        this.logger('info', `Removing listener for ${key} event`, {
+          tags: ['event', 'client'],
+        });
         this.listeners[key] = this.listeners[key].filter((el) => el !== callback);
       },
     };
@@ -868,7 +886,9 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
       this.listeners[key] = [];
     }
 
-    this.logger('info', `Removing listener for ${key} event`, { tags: ['event', 'client'] });
+    this.logger('info', `Removing listener for ${key} event`, {
+      tags: ['event', 'client'],
+    });
     this.listeners[key] = this.listeners[key].filter((value) => value !== callback);
   }
 
@@ -908,7 +928,9 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
     type: string,
     url: string,
     data?: unknown,
-    options: AxiosRequestConfig & { config?: AxiosRequestConfig & { maxBodyLength?: number } } = {},
+    options: AxiosRequestConfig & {
+      config?: AxiosRequestConfig & { maxBodyLength?: number };
+    } = {},
   ): Promise<T> => {
     await this.tokenManager.tokenReady();
     const requestConfig = this._enrichAxiosOptions(options);
@@ -1196,7 +1218,7 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
     }
 
     if (event.channel && event.type === 'notification.message_new') {
-      this.configs[event.channel.type] = event.channel.config;
+      this._addChannelConfig(event.channel);
     }
 
     if (event.type === 'notification.channel_mutes_updated' && event.me?.channel_mutes) {
@@ -1299,9 +1321,13 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
       );
 
       this.logger('info', 'client:recoverState() - Querying channels finished', { tags: ['connection', 'client'] });
-      this.dispatchEvent({ type: 'connection.recovered' } as Event<StreamChatGenerics>);
+      this.dispatchEvent({
+        type: 'connection.recovered',
+      } as Event<StreamChatGenerics>);
     } else {
-      this.dispatchEvent({ type: 'connection.recovered' } as Event<StreamChatGenerics>);
+      this.dispatchEvent({
+        type: 'connection.recovered',
+      } as Event<StreamChatGenerics>);
     }
 
     this.wsPromise = Promise.resolve();
@@ -1331,7 +1357,9 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
       ((this.options.wsConnection as unknown) as StableWSConnection<StreamChatGenerics>).setClient(this);
       this.wsConnection = (this.options.wsConnection as unknown) as StableWSConnection<StreamChatGenerics>;
     } else {
-      this.wsConnection = new StableWSConnection<StreamChatGenerics>({ client: this });
+      this.wsConnection = new StableWSConnection<StreamChatGenerics>({
+        client: this,
+      });
     }
 
     try {
@@ -1353,7 +1381,9 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
 
         this.wsConnection._destroyCurrentWSConnection();
         this.wsConnection.disconnect().then(); // close WS so no retry
-        this.wsFallback = new WSConnectionFallback<StreamChatGenerics>({ client: this });
+        this.wsFallback = new WSConnectionFallback<StreamChatGenerics>({
+          client: this,
+        });
         return await this.wsFallback.connect();
       }
 
@@ -1371,7 +1401,11 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
     const opts = { headers: { 'x-client-request-id': client_request_id } };
     this.doAxiosRequest('get', this.baseURL + '/hi', null, opts).catch((e) => {
       if (this.options.enableInsights) {
-        postInsights('http_hi_failed', { api_key: this.key, err: e, client_request_id });
+        postInsights('http_hi_failed', {
+          api_key: this.key,
+          err: e,
+          client_request_id,
+        });
       }
     });
   }
@@ -1477,7 +1511,11 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
     options: ChannelOptions = {},
     stateOptions: ChannelStateOptions = {},
   ) {
-    const defaultOptions: ChannelOptions = { state: true, watch: true, presence: false };
+    const defaultOptions: ChannelOptions = {
+      state: true,
+      watch: true,
+      presence: false,
+    };
 
     // Make sure we wait for the connect promise if there is a pending one
     await this.wsPromise;
@@ -1513,7 +1551,7 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
     const { skipInitialization, offlineMode = false } = stateOptions;
 
     for (const channelState of channelsFromApi) {
-      this._addChannelConfig(channelState);
+      this._addChannelConfig(channelState.channel);
     }
 
     const channels: Channel<StreamChatGenerics>[] = [];
@@ -1593,7 +1631,10 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
    *
    */
   setLocalDevice(device: BaseDeviceFields) {
-    if (this.wsConnection || this.wsFallback) {
+    if (
+      (this.wsConnection?.isConnecting && this.wsPromise) ||
+      ((this.wsConnection?.isHealthy || this.wsFallback?.isHealthy()) && this._hasConnectionID())
+    ) {
       throw new Error('you can only set device before opening a websocket connection');
     }
 
@@ -1670,8 +1711,8 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
     });
   }
 
-  _addChannelConfig(channelState: ChannelAPIResponse<StreamChatGenerics>) {
-    this.configs[channelState.channel.type] = channelState.channel.config;
+  _addChannelConfig({ cid, config }: ChannelResponse<StreamChatGenerics>) {
+    this.configs[cid] = config;
   }
 
   /**
@@ -1844,10 +1885,11 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
       userMap[userObject.id] = userObject;
     }
 
-    return await this.post<APIResponse & { users: { [key: string]: UserResponse<StreamChatGenerics> } }>(
-      this.baseURL + '/users',
-      { users: userMap },
-    );
+    return await this.post<
+      APIResponse & {
+        users: { [key: string]: UserResponse<StreamChatGenerics> };
+      }
+    >(this.baseURL + '/users', { users: userMap });
   }
 
   /**
@@ -1895,10 +1937,11 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
       }
     }
 
-    return await this.patch<APIResponse & { users: { [key: string]: UserResponse<StreamChatGenerics> } }>(
-      this.baseURL + '/users',
-      { users },
-    );
+    return await this.patch<
+      APIResponse & {
+        users: { [key: string]: UserResponse<StreamChatGenerics> };
+      }
+    >(this.baseURL + '/users', { users });
   }
 
   async deleteUser(
@@ -1909,27 +1952,78 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
       mark_messages_deleted?: boolean;
     },
   ) {
-    return await this.delete<APIResponse & { user: UserResponse<StreamChatGenerics> } & { task_id?: string }>(
-      this.baseURL + `/users/${userID}`,
-      params,
-    );
+    return await this.delete<
+      APIResponse & { user: UserResponse<StreamChatGenerics> } & {
+        task_id?: string;
+      }
+    >(this.baseURL + `/users/${userID}`, params);
   }
 
-  async reactivateUser(
-    userID: string,
-    options?: { created_by_id?: string; name?: string; restore_messages?: boolean },
-  ) {
+  /**
+   * restoreUsers - Restore soft deleted users
+   *
+   * @param {string[]} user_ids which users to restore
+   *
+   * @return {APIResponse} An API response
+   */
+  async restoreUsers(user_ids: string[]) {
+    return await this.post<APIResponse>(this.baseURL + `/users/restore`, {
+      user_ids,
+    });
+  }
+
+  /**
+   * reactivateUser - Reactivate one user
+   *
+   * @param {string} userID which user to reactivate
+   * @param {ReactivateUserOptions} [options]
+   *
+   * @return {UserResponse} Reactivated user
+   */
+  async reactivateUser(userID: string, options?: ReactivateUserOptions) {
     return await this.post<APIResponse & { user: UserResponse<StreamChatGenerics> }>(
       this.baseURL + `/users/${userID}/reactivate`,
       { ...options },
     );
   }
 
-  async deactivateUser(userID: string, options?: { created_by_id?: string; mark_messages_deleted?: boolean }) {
+  /**
+   * reactivateUsers - Reactivate many users asynchronously
+   *
+   * @param {string[]} user_ids which users to reactivate
+   * @param {ReactivateUsersOptions} [options]
+   *
+   * @return {TaskResponse} A task ID
+   */
+  async reactivateUsers(user_ids: string[], options?: ReactivateUsersOptions) {
+    return await this.post<APIResponse & TaskResponse>(this.baseURL + `/users/reactivate`, { user_ids, ...options });
+  }
+
+  /**
+   * deactivateUser - Deactivate one user
+   *
+   * @param {string} userID which user to deactivate
+   * @param {DeactivateUsersOptions} [options]
+   *
+   * @return {UserResponse} Deactivated user
+   */
+  async deactivateUser(userID: string, options?: DeactivateUsersOptions) {
     return await this.post<APIResponse & { user: UserResponse<StreamChatGenerics> }>(
       this.baseURL + `/users/${userID}/deactivate`,
       { ...options },
     );
+  }
+
+  /**
+   * deactivateUsers - Deactivate many users asynchronously
+   *
+   * @param {string[]} user_ids which users to deactivate
+   * @param {DeactivateUsersOptions} [options]
+   *
+   * @return {TaskResponse} A task ID
+   */
+  async deactivateUsers(user_ids: string[], options?: DeactivateUsersOptions) {
+    return await this.post<APIResponse & TaskResponse>(this.baseURL + `/users/deactivate`, { user_ids, ...options });
   }
 
   async exportUser(userID: string, options?: Record<string, string>) {
@@ -2339,7 +2433,9 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
     );
     return this.partialUpdateMessage(
       messageId,
-      ({ set: { pinned: false } } as unknown) as PartialMessageUpdate<StreamChatGenerics>,
+      ({
+        set: { pinned: false },
+      } as unknown) as PartialMessageUpdate<StreamChatGenerics>,
       userId,
     );
   }
@@ -2390,7 +2486,9 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
       if (isString(userId)) {
         clonedMessage.user_id = userId;
       } else {
-        clonedMessage.user = { id: userId.id } as UserResponse<StreamChatGenerics>;
+        clonedMessage.user = {
+          id: userId.id,
+        } as UserResponse<StreamChatGenerics>;
       }
     }
 
@@ -2480,6 +2578,11 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
   ): AxiosRequestConfig {
     const token = this._getToken();
     const authorization = token ? { Authorization: token } : undefined;
+    let signal: AbortSignal | null = null;
+    if (this.nextRequestAbortController !== null) {
+      signal = this.nextRequestAbortController.signal;
+      this.nextRequestAbortController = null;
+    }
 
     if (!options.headers?.['x-client-request-id']) {
       options.headers = {
@@ -2501,6 +2604,7 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
         'X-Stream-Client': this.getUserAgent(),
         ...options.headers,
       },
+      ...(signal ? { signal } : {}),
       ...options.config,
     };
   }
@@ -2557,7 +2661,9 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
    * @returns {Promise<APIResponse>}
    */
   createPermission(permissionData: CustomPermissionOptions) {
-    return this.post<APIResponse>(`${this.baseURL}/permissions`, { ...permissionData });
+    return this.post<APIResponse>(`${this.baseURL}/permissions`, {
+      ...permissionData,
+    });
   }
 
   /** updatePermission - updates an existing custom permission
@@ -2567,7 +2673,9 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
    * @returns {Promise<APIResponse>}
    */
   updatePermission(id: string, permissionData: Omit<CustomPermissionOptions, 'id'>) {
-    return this.put<APIResponse>(`${this.baseURL}/permissions/${id}`, { ...permissionData });
+    return this.put<APIResponse>(`${this.baseURL}/permissions/${id}`, {
+      ...permissionData,
+    });
   }
 
   /** deletePermission - deletes a custom permission
@@ -2906,7 +3014,7 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
    * @param {string[]} user_ids which users to delete
    * @param {DeleteUserOptions} options Configuration how to delete users
    *
-   * @return {APIResponse} A task ID
+   * @return {TaskResponse} A task ID
    */
   async deleteUsers(user_ids: string[], options: DeleteUserOptions) {
     if (options?.user !== 'soft' && options?.user !== 'hard') {
@@ -3029,5 +3137,22 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
    */
   async listPushProviders() {
     return await this.get<APIResponse & PushProviderListResponse>(this.baseURL + `/push_providers`);
+  }
+
+  /**
+   * creates an abort controller that will be used by the next HTTP Request.
+   */
+  createAbortControllerForNextRequest() {
+    return (this.nextRequestAbortController = new AbortController());
+  }
+
+  /**
+   * commits a pending message, making it visible in the channel and for other users
+   * @param id the message id
+   *
+   * @return {APIResponse & MessageResponse} The message
+   */
+  async commitMessage(id: string) {
+    return await this.post<APIResponse & MessageResponse>(this.baseURL + `/messages/${id}/commit`);
   }
 }