+++ test/unit/utils.js
@@ -1,7 +1,24 @@
 import chai from 'chai';
-import { normalizeQuerySort } from '../../src/utils';
+import { generateUUIDv4, normalizeQuerySort } from '../../src/utils';
+import sinon from 'sinon';
 
 const expect = chai.expect;
+
+describe('generateUUIDv4', () => {
+	beforeEach(() => {
+		sinon.restore();
+	});
+
+	it('generates a UUID manually when crypto is unavailable', () => {
+		sinon.spy(Math, 'pow');
+		sinon.spy(Math, 'random');
+		const uuid = generateUUIDv4();
+		expect(uuid).to.be.a('string');
+		expect(uuid.length).to.equal(36);
+		expect(Math.pow.calledWithMatch(2, 8)).to.be.true;
+		expect(Math.random.callCount).to.be.equal(16);
+	});
+});
 
 describe('test if sort is deterministic', () => {
 	it('test sort object', () => {